requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==4.1.1
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.24.2
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
zope.interface==5.4.0
